% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scXCL_gui.R
\name{scXCL_vis}
\alias{scXCL_vis}
\title{Visualize scXCL result in a web browser.}
\usage{
scXCL_vis(XCL_result)
}
\arguments{
\item{XCL_result}{an object generate by scXCL funtion}
}
\description{
Runs interactive \code{shiny} session of \code{scXCL}
}
